backend/requirements/base.txt
# DisplayDeck Backend Requirements

# Base requirements for all environments
django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-channels>=4.0.0
channels-redis>=4.1.0
django-guardian>=2.4.0
psycopg2-binary>=2.9.7
redis>=5.0.0
celery>=5.3.0
pillow>=10.0.0
python-decouple>=3.8
whitenoise>=6.5.0

# API Documentation
drf-spectacular>=0.26.0

# File handling and storage
django-storages>=1.14.0
boto3>=1.28.0  # For AWS S3 if needed

# Security
cryptography>=41.0.0